Set Up a Django Application for Stock Monitoring
I have created the necessary template files for our stock application, which will define the user interface. Additionally, I have created a separate file called urls.py specifically for handling the URLs related to the stock application.

To ensure that the URLs of the stock application are accessible, I included an import statement in the main application's urls.py file. By doing this, I ensured proper routing and mapping of URLs to the corresponding views within the stock application.